240507 임재현 DB 구조 변경 사항 적용

+++ server/src/main/java/org/web/application/server/entity/HobbyEntity.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class HobbyEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long hobbyId;
+    private Integer hobbyId;
 
     private String hobbyName;
